+++ src/main/java/net/outmoded/outmodedlib/packer/jsonObjects/unicode/providers/UnihexSizeOverrides.java
@@ -1,7 +1,7 @@
 package net.outmoded.outmodedlib.packer.jsonObjects.unicode.providers;
 
-public record UnihexSizeOverrides(String from, String to, int left, int right){
-    public UnihexSizeOverrides(String from, String to, int left, int right){
+public record UnihexSizeOverrides(String from, String to, double left, double right){
+    public UnihexSizeOverrides(String from, String to, double left, double right){
         this.from = from;
         this.to = to;
         this.left = left;
